Simplify processInput method
The use of graphs and trees to deal with the processing of inputs is to hard to wrap my head around and makes scripting complicated, especially when dealing with room and game state dependant responses.

Replace that with simple conditional statements for the moment. I'll optimize later, when I have a better understanding of the requirements.